#25 Refactoring and improvement of the tests to the new design of the activity. The previous tests were simply instantiating the fragments, now the events are being simulated in order to produce a real scenario for the tests to be performed.
The ExperimentSensorSetupFragment test is failing due to the method
onCreate of the fragment ExperimentSensorListSelection do not be called
during the execution of the tests (it happens only when running the
application), and so the list of sensors from which one sensor needs to
be selected to get into the fragment being tested, is not being loaded,
denying so the test to reach the fragment.

The method lView.performItemClick is being used to perform the click
that triggers the event responsible by creating the fragments.