#129 Fixed the app crashing on API level 8 The app crashing on API level 8 was being caused due to the fact that the support library does not have support to the ArrayAdapter::addAll method, which is used to set the list of experiments on the ListView.
An implementation of the ArrayAdapter identical to the one from the
Android framework was added in the project, so that the level 8 APIs can
use it. Other possible solution would be to add the items using a for
loop, but this approach lose in performance compared to the one used,
once that it is using the native Java collections procedure.